moved OC handler to its own module, and nits.
