+++ runtime/handlers/operation_container/operation_container_handler_test.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"log/slog"
 	"strings"
+	"testing"
 
 	oc "github.com/Azure/OperationContainer/api/v1"
 	ocMock "github.com/Azure/OperationContainer/api/v1/mock"
@@ -20,7 +21,12 @@ import (
 	"go.uber.org/mock/gomock"
 )
 
-var _ = Describe("OperationContainerHandler Testing", func() {
+func TestQoSErrorHandler(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "OperationContainerHandler Suite")
+}
+
+var _ = Describe("OperationContainerHandler", func() {
 	var (
 		ctrl                     *gomock.Controller
 		ctx                      context.Context
@@ -202,3 +208,36 @@ var _ = Describe("OperationContainerHandler Testing", func() {
 		})
 	})
 })
+
+func SampleHandler() shuttle.HandlerFunc {
+	return func(ctx context.Context, settler shuttle.MessageSettler, message *azservicebus.ReceivedMessage) {
+	}
+}
+
+func SampleErrorHandler(testErrorMessage error) handlerErrors.ErrorHandlerFunc {
+	return func(ctx context.Context, settler shuttle.MessageSettler, message *azservicebus.ReceivedMessage) error {
+		return testErrorMessage
+	}
+}
+
+type fakeMessageSettler struct{}
+
+func (f *fakeMessageSettler) AbandonMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.AbandonMessageOptions) error {
+	return nil
+}
+func (f *fakeMessageSettler) CompleteMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.CompleteMessageOptions) error {
+	failureMessage := "failure_test"
+	if message.ContentType != nil && strings.Compare(*message.ContentType, failureMessage) == 0 {
+		return errors.New("settler error")
+	}
+	return nil
+}
+func (f *fakeMessageSettler) DeadLetterMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.DeadLetterOptions) error {
+	return nil
+}
+func (f *fakeMessageSettler) DeferMessage(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.DeferMessageOptions) error {
+	return nil
+}
+func (f *fakeMessageSettler) RenewMessageLock(ctx context.Context, message *azservicebus.ReceivedMessage, options *azservicebus.RenewMessageLockOptions) error {
+	return nil
+}
